requirements.txt
streamlit
openai
requests
pandas
Pillow
python-dotenv 
